// src/types/entities.ts
export type Period = {
    location: string;
    from: string;
    to?: string;
    title: string;
    desc: string;
    link?: string;
    link_name: string;
    link1?:string;
    link1_name?:string;
    link2?:string;
    link2_name?:string;
    link3?:string;
    link3_name?:string;
};

export type Document = {
    title: string;
    link: string;
};